Organize desk stack behavior
