+++ apps/ui/src/ui-desks/desk/provider/index.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { StackAnimationProvider } from '@/ui-desks/stacks/context';
-import { getStackId, isStackExpanded } from '@/ui-desks/stacks/utils';
+import { StackProvider } from '@/ui-desks/stacks/provider';
 import {
 	DeskContext,
 	DeskEditorRegistrationContext,
@@ -10,9 +9,7 @@ import {
 } from './context';
 import {
 	addWidgetToEditor,
-	collapseAllExpandedStacksInEditor,
 	createDeskConfigFromEditor,
-	expandStackInEditor,
 	getCurrentSelectedWidgetToolbarItem,
 	hasCameraChange,
 	hydrateEditorFromDesk,
@@ -22,7 +19,7 @@ import {
 	updateSelectedWidgetPropsInEditor,
 } from './editor-state';
 import { useDeskPersistence } from './persistence';
-import type { Editor, TLEventInfo } from 'tldraw';
+import type { Editor } from 'tldraw';
 
 export { useDesk, useRegisterDeskEditor } from './context';
 
@@ -96,118 +93,6 @@ export function DeskProvider( { siteId, children }: DeskProviderProps ) {
 			unsubscribeSession();
 		};
 	}, [ editor, saveDeskConfig ] );
-
-	useEffect( () => {
-		if ( ! editor ) {
-			return;
-		}
-
-		const selectStackMembersForDrag = ( info: TLEventInfo ) => {
-			if ( info.type !== 'pointer' || info.name !== 'pointer_down' || info.button !== 0 ) {
-				return;
-			}
-
-			const stackId = getCollapsedStackIdAtPointer( editor );
-			if ( ! stackId ) {
-				return;
-			}
-
-			const memberIds = editor
-				.getCurrentPageShapes()
-				.filter( ( shape ) => getStackId( shape ) === stackId )
-				.map( ( shape ) => shape.id );
-			if ( memberIds.length > 1 ) {
-				editor.setSelectedShapes( memberIds );
-			}
-		};
-
-		editor.on( 'event', selectStackMembersForDrag );
-		return () => {
-			editor.off( 'event', selectStackMembersForDrag );
-		};
-	}, [ editor ] );
-
-	useEffect( () => {
-		if ( ! editor ) {
-			return;
-		}
-
-		let isPointerSession = false;
-		let pointerDownStackId: string | null = null;
-		const movedShapeIds = new Set< string >();
-		const unsubscribeShapeChanges = editor.sideEffects.registerAfterChangeHandler(
-			'shape',
-			( previousShape, nextShape ) => {
-				if ( ! isPointerSession ) {
-					return;
-				}
-
-				if (
-					previousShape.x !== nextShape.x ||
-					previousShape.y !== nextShape.y ||
-					previousShape.rotation !== nextShape.rotation
-				) {
-					movedShapeIds.add( nextShape.id );
-				}
-			}
-		);
-
-		const handleStackClick = ( info: TLEventInfo ) => {
-			if ( info.type !== 'pointer' ) {
-				return;
-			}
-
-			if ( info.name === 'pointer_down' ) {
-				if ( info.button !== 0 ) {
-					return;
-				}
-
-				isPointerSession = true;
-				pointerDownStackId = getCollapsedStackIdAtPointer( editor );
-				movedShapeIds.clear();
-				return;
-			}
-
-			if ( info.name !== 'pointer_up' ) {
-				return;
-			}
-			if ( ! isPointerSession ) {
-				return;
-			}
-
-			isPointerSession = false;
-			const clickedStackId = pointerDownStackId;
-			pointerDownStackId = null;
-			if ( clickedStackId ) {
-				const movedStack = editor
-					.getCurrentPageShapes()
-					.filter( ( shape ) => getStackId( shape ) === clickedStackId )
-					.some( ( shape ) => movedShapeIds.has( shape.id ) );
-				movedShapeIds.clear();
-				if ( movedStack ) {
-					return;
-				}
-				if ( expandStackInEditor( editor, clickedStackId ) ) {
-					editor.setSelectedShapes( [] );
-				}
-				return;
-			}
-
-			const selectedShapeIds = editor.getSelectedShapeIds();
-			if ( selectedShapeIds.length === 0 ) {
-				collapseAllExpandedStacksInEditor( editor );
-				movedShapeIds.clear();
-				return;
-			}
-			movedShapeIds.clear();
-		};
-
-		editor.on( 'event', handleStackClick );
-		return () => {
-			editor.off( 'event', handleStackClick );
-			unsubscribeShapeChanges();
-		};
-	}, [ editor ] );
 
 	const registerEditor = useCallback( ( nextEditor: Editor | null ) => {
 		setEditor( nextEditor );
@@ -309,18 +194,8 @@ export function DeskProvider( { siteId, children }: DeskProviderProps ) {
 	return (
 		<DeskEditorRegistrationContext.Provider value={ registerEditor }>
 			<DeskContext.Provider value={ value }>
-				<StackAnimationProvider>{ children }</StackAnimationProvider>
+				<StackProvider editor={ editor }>{ children }</StackProvider>
 			</DeskContext.Provider>
 		</DeskEditorRegistrationContext.Provider>
 	);
-}
-
-function getCollapsedStackIdAtPointer( editor: Editor ) {
-	const hitShape = editor.getShapeAtPoint( editor.inputs.currentPagePoint, {
-		hitInside: true,
-		renderingOnly: true,
-		margin: editor.options.hitTestMargin / editor.getZoomLevel(),
-	} );
-	const stackId = getStackId( hitShape );
-	return stackId && ! isStackExpanded( hitShape ) ? stackId : null;
 }
