+++ apps/ui/src/ui-desks/desk/tldraw-hover-state-sync.tsx
@@ -1,0 +1,44 @@
+import { useEffect } from 'react';
+import { type TLEventInfo, useEditor } from 'tldraw';
+
+export function TldrawHoverStateSync() {
+	const editor = useEditor();
+
+	useEffect( () => {
+		const markHoveringCanvas = ( info: TLEventInfo ) => {
+			if ( info.type !== 'pointer' || info.name !== 'pointer_move' ) {
+				return;
+			}
+
+			const instanceState = editor.getInstanceState();
+			if ( instanceState.isCoarsePointer || instanceState.isHoveringCanvas === true ) {
+				return;
+			}
+
+			editor.updateInstanceState( { isHoveringCanvas: true } );
+		};
+		const markNotHoveringCanvas = () => {
+			if ( editor.getInstanceState().isHoveringCanvas === false ) {
+				return;
+			}
+
+			editor.updateInstanceState( { isHoveringCanvas: false } );
+		};
+		const handleVisibilityChange = () => {
+			if ( document.visibilityState === 'hidden' ) {
+				markNotHoveringCanvas();
+			}
+		};
+
+		editor.on( 'event', markHoveringCanvas );
+		window.addEventListener( 'blur', markNotHoveringCanvas );
+		document.addEventListener( 'visibilitychange', handleVisibilityChange );
+		return () => {
+			editor.off( 'event', markHoveringCanvas );
+			window.removeEventListener( 'blur', markNotHoveringCanvas );
+			document.removeEventListener( 'visibilitychange', handleVisibilityChange );
+		};
+	}, [ editor ] );
+
+	return null;
+}
