+++ apps/ui/src/ui-desks/shapes/rectangle-widget/shape-util.tsx
@@ -6,13 +6,12 @@ import {
 	resizeBox,
 	useEditor,
 	useIsEditing,
-	useValue,
 	type JsonObject,
 	type RecordProps,
 	type TLResizeInfo,
 } from 'tldraw';
-import { useStackAnimation } from '@/ui-desks/stacks/context';
-import { getStackId, getStackOrder, isStackExpanded } from '@/ui-desks/stacks/utils';
+import { useStackShapeInteraction } from '@/ui-desks/stacks/use-stack-shape-interaction';
+import { getStackId, isStackExpanded } from '@/ui-desks/stacks/utils';
 import { getWidgetDefinition } from '@/ui-desks/widgets/registry';
 import {
 	RECTANGLE_WIDGET_SHAPE_TYPE,
@@ -141,7 +140,7 @@ function RectangleWidgetComponent( {
 } ) {
 	const editor = useEditor();
 	const isEditing = useIsEditing( shape.id );
-	const stackInteraction = useStackInteraction( shape );
+	const stackInteraction = useStackShapeInteraction( shape );
 	const WidgetComponent = definition.Component as unknown as ComponentType<
 		DeskWidgetComponentProps< JsonObject >
 	>;
@@ -167,48 +166,6 @@ function RectangleWidgetComponent( {
 			/>
 		</div>
 	);
-}
-
-function useStackInteraction( shape: RectangleWidgetShape ) {
-	const editor = useEditor();
-	const { pressedStackId, pressStack } = useStackAnimation();
-	const stackId = getStackId( shape );
-	const isExpanded = isStackExpanded( shape );
-	const hoveredStackId = useValue(
-		'desk-stack-hovered-stack-id',
-		() => {
-			const hoveredShape = editor.getHoveredShape();
-			const hoveredShapeStackId = getStackId( hoveredShape );
-			return hoveredShapeStackId && ! isStackExpanded( hoveredShape ) ? hoveredShapeStackId : null;
-		},
-		[ editor ]
-	);
-	const isHovered = Boolean( stackId ) && ! isExpanded && hoveredStackId === stackId;
-	const isPressed = Boolean( stackId ) && ! isExpanded && pressedStackId === stackId;
-	const order = getStackOrder( shape );
-	const members = stackId
-		? editor.getCurrentPageShapes().filter( ( member ) => getStackId( member ) === stackId )
-		: [];
-	const center = ( members.length - 1 ) / 2;
-	const step = order - center;
-	const hoverTranslate = isHovered ? step * 7 : 0;
-	const hoverRotate = isHovered ? step * 2.5 : 0;
-	const pressScale = isPressed ? 0.985 : 1;
-
-	return {
-		style: {
-			width: '100%',
-			height: '100%',
-			transform: `translate(${ hoverTranslate }px, ${ hoverTranslate }px) rotate(${ hoverRotate }deg) scale(${ pressScale })`,
-			transformOrigin: 'center',
-			transition: 'transform 220ms ease',
-		},
-		onPointerDown: () => {
-			if ( stackId && ! isExpanded ) {
-				pressStack( stackId );
-			}
-		},
-	};
 }
 
 function getWidgetIdFromShapeId( shapeId: string ) {
